Add now defunct web scraper/encounter tracker
This program scrapes data to act as a companion tool for a playthrough of Pokemon Black & White, using a special ruleset for added difficulty. While the web scraper no longer works due to a reformat of Bulbapedia's (the source from which this program scrapes) HTML, if one loads the elevationlocke user profile, there exists the complete scraped and reformatted data for a working companion tool for a playthrough of Pokemon Black.